Don’t copy webp.framework into SDL2_image.framework
This means WebP.framework needs to linked and copied in other projects wherever SDL2_image.framework is, but this was easier than getting it to load from the right place and simplifies the build.

Also fixes the WebP which I broke somewhere in the last two builds :/ although there is still something weird in SDL2_image - if built from clean it fails, if built after SDL2 has been build by itself its fine (left a note at the end of the build).